tests/submission/W3C/domhtml/HTMLImageElement10.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IMG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-img-element.html#dom-img-usemap" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<img align="middle" alt="DTS IMAGE LOGO" border="0" height="50" hspace="4" id="IMAGE-1" longdesc="./files/desc.html" name="IMAGE-1" src="/support/images/yellow-rectangle.png" usemap="#DTS-MAP" vspace="10" width="100">
</p>



<script>
var domtest = async_test("The useMap attribute specifies to use the client-side image map.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vusemap;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("img");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vusemap = testNode.useMap;

      assert_equals(vusemap,"#DTS-MAP");
       
    domtest.done();
}));

</script>
<p>The useMap attribute specifies to use the client-side image map.

    Retrieve the useMap attribute and examine its value.</p>
<div id="log"></div>
</body></html>
